backend/requirements.txt
fastapi==0.115.5
uvicorn==0.31.0
pydantic==2.9.2
pymongo==4.8.0
python-dotenv==1.0.1
email-validator==2.2.0